+++ lockfile_generator/src/gradle.rs
@@ -1,5 +1,6 @@
 //! Java gradle ecosystem.
 
+use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 
@@ -23,5 +24,15 @@ impl Generator for Gradle {
 
     fn tool(&self) -> &'static str {
         "Gradle"
+    }
+
+    fn check_prerequisites(&self, manifest_path: &Path) -> Result<()> {
+        if manifest_path.file_name() == Some(OsStr::new("build.gradle"))
+            || manifest_path.file_name() == Some(OsStr::new("build.gradle.kts"))
+        {
+            Ok(())
+        } else {
+            Err(Error::InvalidManifest(manifest_path.to_path_buf()))
+        }
     }
 }